ProjectManagement
Folders and to-do lists for future projects and work needed to be done. Can be further organized in the future but notes and other important information can be put here.

Currently everything is blank and basically a placeholder but we should use either github or another service to maintain documentation and track of what we are doing and what needs to be done as well as what has been done.

I doubt you will be reading this but the point still stands, I look forward to future work!